+++ packages/manager/apps/cloud/client/app/vps/header/vps-header.controller.js
@@ -1,10 +1,16 @@
 import includes from 'lodash/includes';
+import filter from 'lodash/filter';
+import map from 'lodash/map';
+import snakeCase from 'lodash/snakeCase';
+
+import {
+  TAB_FEATURES,
+} from './vps-header.constants';
 
 export default class {
   /* @ngInject */
   constructor(
     $rootScope,
-    $stateParams,
     $translate,
     CucCloudMessage,
     CucFeatureAvailabilityService,
@@ -15,17 +21,16 @@ export default class {
     STOP_NOTIFICATION_USER_PREF,
   ) {
     this.$rootScope = $rootScope;
-    this.$stateParams = $stateParams;
     this.$translate = $translate;
     this.CucCloudMessage = CucCloudMessage;
     this.CucFeatureAvailabilityService = CucFeatureAvailabilityService;
+    this.CucProductsService = CucProductsService;
     this.OvhApiMe = OvhApiMe;
     this.VpsNotificationIpv6 = VpsNotificationIpv6;
-    this.STOP_NOTIFICATION_USER_PREF = STOP_NOTIFICATION_USER_PREF;
-    this.serviceName = $stateParams.serviceName;
-    this.description = $stateParams.serviceName;
     this.VpsService = VpsService;
-    this.CucProductsService = CucProductsService;
+    this.STOP_NOTIFICATION_USER_PREF = STOP_NOTIFICATION_USER_PREF;
+
+    this.description = this.serviceName;
 
     this.loaders = {
       init: false,
@@ -62,6 +67,23 @@ export default class {
       .catch(() => this.CucCloudMessage.error(this.$translate.instant('vps_dashboard_loading_error')))
       .finally(() => { this.loaders.init = false; });
     this.description = this.CucProductsService.getDisplayName('VPS', this.serviceName);
+    this.features = this.getFeatures();
+    [this.feature] = this.features;
+  }
+
+  getFeatures() {
+    return map(
+      filter(
+        map(
+          TAB_FEATURES, 'title',
+        ),
+        feature => this.capabilities.includes(feature),
+      ),
+      feature => ({
+        textId: `vps_tab_${snakeCase(feature)}`,
+        state: `iaas.vps.detail.${snakeCase(feature)}`,
+      }),
+    );
   }
 
   checkMessages(vps) {