+++ packages/manager/apps/dedicated/client/app/app.js
@@ -28,12 +28,15 @@ import ovhContacts from '@ovh-ux/ng-ovh-contacts';
 import ovhManagerCore from '@ovh-ux/manager-core';
 import ovhManagerBanner from '@ovh-ux/manager-banner';
 import ovhManagerEnterpriseCloudDatabase from '@ovh-ux/manager-enterprise-cloud-database';
+import ovhManagerMfaEnrollment from '@ovh-ux/mfa-enrollment';
 import ovhManagerNavbar from '@ovh-ux/manager-navbar';
 import ovhManagerServerSidebar from '@ovh-ux/manager-server-sidebar';
+import ovhManagerSupport from '@ovh-ux/manager-support';
 import ovhPaymentMethod from '@ovh-ux/ng-ovh-payment-method';
 import uiRouter, { RejectType } from '@uirouter/angularjs';
 
 import moduleExchange from '@ovh-ux/manager-exchange';
+import account from './account';
 import config from './config/config';
 import contactsService from './account/contacts/service/contacts-service.module';
 import dedicatedCloudDatacenterDrp from './dedicatedCloud/datacenter/drp';
@@ -46,7 +49,6 @@ import preload from './components/manager-preload/manager-preload.module';
 import dedicatedServerBandwidth from './dedicated/server/bandwidth/bandwidth.module';
 import dedicatedServerInterfaces from './dedicated/server/interfaces/interfaces.module';
 import dedicatedServerServers from './dedicated/server/servers/servers.module';
-import mfaEnrollment from './mfaEnrollment/mfaEnrollment.module';
 
 
 Environment.setRegion(__WEBPACK_REGION__);
@@ -55,6 +57,7 @@ Environment.setVersion(__VERSION__);
 angular
   .module('App', [
     __NG_APP_INJECTIONS__,
+    account,
     ovhManagerCore,
     'Billing',
     'chart.js',
@@ -74,7 +77,7 @@ angular
     'Module.ip',
     'Module.license',
     'Module.otrs',
-    mfaEnrollment,
+    ovhManagerMfaEnrollment,
     'ng.ckeditor',
     'ngMessages',
     ngAtInternet,
@@ -104,6 +107,7 @@ angular
     ovhManagerPccDashboard,
     ovhManagerPccResourceUpgrade,
     ovhManagerServerSidebar,
+    ovhManagerSupport,
     ngTailLogs,
     'ovhBrowserAlert',
     ovhContacts,