+++ packages/manager/apps/public-cloud/src/index.js
@@ -20,13 +20,12 @@ import has from 'lodash/has';
 
 import navbar from '@ovh-ux/manager-navbar';
 import ovhManagerCore from '@ovh-ux/manager-core';
+import ovhManagerMfaEnrollment from '@ovh-ux/mfa-enrollment';
 import ovhManagerPci from '@ovh-ux/manager-pci';
 import ngOvhApiWrappers from '@ovh-ux/ng-ovh-api-wrappers';
-import ngOvhOtrs from '@ovh-ux/ng-ovh-otrs';
 import ngOvhUserPref from '@ovh-ux/ng-ovh-user-pref';
 import ngUiRouterBreadcrumb from '@ovh-ux/ng-ui-router-breadcrumb';
 import ngUiRouterLineProgress from '@ovh-ux/ng-ui-router-line-progress';
-import { OTRS_URLS } from './index.constants';
 
 import 'ovh-ui-kit/dist/oui.css';
 import 'ovh-ui-kit-bs/dist/ovh-ui-kit-bs.css';
@@ -53,9 +52,9 @@ angular
     ngUiRouterBreadcrumb,
     ngUiRouterLineProgress,
     ovhManagerCore,
+    ovhManagerMfaEnrollment,
     ovhManagerPci,
     ngOvhApiWrappers,
-    ngOvhOtrs,
     ngOvhUserPref,
     navbar,
     preload,
@@ -70,9 +69,6 @@ angular
     const defaultLanguage = TranslateServiceProvider.getUserLocale();
     // set moment locale
     moment.locale(defaultLanguage.split('_')[0]);
-  })
-  .config((coreConfigProvider, OtrsPopupProvider) => {
-    OtrsPopupProvider.setBaseUrlTickets(OTRS_URLS[coreConfigProvider.region]);
   })
   .run(/* @ngInject */ ($state) => {
     $state.defaultErrorHandler((error) => {