+++ packages/manager/modules/mfa-enrollment/src/mfa-enrollment.component.js
@@ -1,0 +1,11 @@
+import controller from './mfa-enrollment.controller';
+import template from './mfa-enrollment.html';
+
+export default {
+  bindings: {
+    from: '<',
+    rootState: '<',
+  },
+  controller,
+  template,
+};