+++ packages/manager/modules/pci/src/projects/new/new.controller.js
@@ -105,7 +105,6 @@ export default class PciProjectNewCtrl {
   isPaymentMethodIntegrationVisible() {
     const currentStep = this.getCurrentStep();
 
-
     return currentStep.name !== 'description'
       && !this.paymentStatus()
       && !get(currentStep, 'model.defaultPaymentMethod');