+++ packages/manager/modules/pci/src/projects/new/payment/challenge/challenge.controller.js
@@ -24,7 +24,7 @@ export default class PciProjectsNewPaymentChallengeController {
 
   $onInit() {
     this.isChallenging = false;
-    if (get(this.defaultPaymentMethod, 'paymentType.value') === 'CREDIT_CARD') {
+    if (get(this.defaultPaymentMethod, 'paymentType') === 'CREDIT_CARD') {
       this.creditCardSuffix = this.defaultPaymentMethod.label.substring(6);
     }
     this.paypalUrl = BUILD_PAYPAL_URL(PAYPAL_LOGIN_URL, PAYPAL_LOGIN_PARAMS);