+++ packages/manager/modules/support/src/support.routing.js
@@ -1,0 +1,14 @@
+import component from './support.component';
+
+export const state = {
+  name: 'support',
+  redirectTo: 'support.tickets',
+  url: '/support',
+  views: {
+    '@': component.name,
+  },
+};
+
+export default {
+  state,
+};