+++ packages/manager/modules/support/src/tickets/new-ticket/new-ticket.component.js
@@ -1,0 +1,14 @@
+import controller from './new-ticket.controller';
+import template from './new-ticket.html';
+
+export default {
+  bindings: {
+    goToTickets: '&',
+    serviceName: '<?',
+    serviceTypeName: '<?',
+    urls: '<',
+  },
+  controller,
+  name: 'supportNew',
+  template,
+};