+++ packages/manager/apps/pci/CHANGELOG.md
@@ -1,3 +1,12 @@
+# [2.2.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@2.1.4...@ovh-ux/manager-pci-app@2.2.0) (2019-10-28)
+
+
+### Features
+
+* **pci:** replace OTRS by Support ([174c858](https://github.com/ovh/manager/commit/174c858aded60ecfd872a5c92bee735fafe03cf2))
+
+
+
 ## [2.1.4](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@2.1.3...@ovh-ux/manager-pci-app@2.1.4) (2019-10-25)